refactor getKeywordDensity and keyphraseLength Abstract researches
and delete the two researches from English folder
