+++ packages/yoastseo/src/languageProcessing/researches/getKeywordDensity.js
@@ -1,25 +1,25 @@
 /** @module analyses/getKeywordDensity */
 
-import countWords from "../../helpers/word/countWords.js";
+import countWords from "../helpers/word/countWords.js";
 
 /**
  * Calculates the keyword density.
  *
  * @param {Object} paper        The paper containing keyword and text.
  * @param {Object} researcher   The researcher.
- * @param {boolean} hasStemmer  Whether the language has the stemmer or not.
  *
- * @returns {Object} The keyword density and whether there is a stemmer available or not.
+ * @returns {Object} The keyword density.
  */
-export default function( paper, researcher, hasStemmer ) {
+export default function( paper, researcher ) {
 	const wordCount = countWords( paper.getText() );
 	if ( wordCount === 0 ) {
 		return 0;
 	}
 
 	const keywordCount = researcher.getResearch( "keywordCount" );
+	const stemmer = researcher.getHelper( "getStemmer" );
 	return {
 		keywordDensity: ( keywordCount.count / wordCount ) * 100,
-		stemmer: hasStemmer,
+		stemmer: stemmer,
 	};
 }
