+++ packages/yoastseo/src/scoring/assessments/seo/KeywordDensityAssessment.js
@@ -7,6 +7,7 @@ import { inRangeEndInclusive, inRangeStartEndInclusive, inRangeStartInclusive } 
 import { createAnchorOpeningTag } from "../../../helpers/shortlinker";
 import keyphraseLengthFactor from "../../helpers/assessments/keyphraseLengthFactor.js";
 import countWords from "./../../../languageProcessing/helpers/word/countWords";
+import baseStemmer from "../../../languageProcessing/helpers/morphology/baseStemmer";
 
 /**
  * Represents the assessment that will look if the keyphrase density is within the recommended range.
@@ -109,7 +110,7 @@ class KeywordDensityAssessment extends Assessment {
 		this._keywordDensityData = researcher.getResearch( "getKeywordDensity" );
 
 		this._hasMorphologicalForms = researcher.getData( "morphology" ) !== false &&
-			this._keywordDensityData.stemmer === true;
+			this._keywordDensityData.stemmer !== baseStemmer;
 
 		this._keywordDensity = this._keywordDensityData.keywordDensity * keyphraseLengthFactor( keyphraseLength );
 		const calculatedScore = this.calculateResult( i18n );
